@@ -1,8 +1,9 @@
 package com.andrognito.pinlockview;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Created by aritraroy on 31/05/16.
